:goat: Split up memory tests
